+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## UNRELEASED
 
 ### Added
+- Add the operator `apalache::generate` to mirror `Apalache!Gen` (see #1455).
+
 ### Changed
 
 - Bump Apalache to 0.49.0 (including support for `apalache::generate` [#3138](https://github.com/apalache-mc/apalache/pull/3138))